[add] add exception/page_fault_handler.rs
